requirements.txt
autopep8==1.4.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycodestyle==2.5.0
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
